requirements.txt
sanic~=22.3.2
aiomysql~=0.1.1
requests~=2.27.1
PyYAML~=6.0